+++ src/parser/block/dlist.ts
@@ -1,7 +1,6 @@
 import { DListParser } from '../block';
-import { union, inits, some, block, line, validate, rewrite, context, creator, open, convert, trim, lazy, fmap } from '../../combinator';
-import { defrag } from '../util';
-import { blankline } from './paragraph';
+import { union, inits, some, block, line, validate, rewrite, context, creator, open, trim, lazy, fmap } from '../../combinator';
+import { justify, defrag } from '../util';
 import { inline, indexer, indexee } from '../inline';
 import { anyline } from '../source';
 import { push } from 'spica/array';
@@ -9,9 +8,8 @@ import { html } from 'typed-dom';
 
 export const dlist: DListParser = lazy(() => block(fmap(validate(
   /^~[^\S\n]/,
-  convert(source => source.replace(blankline, ''),
   context({ syntax: { inline: { media: false } } },
-  some(inits([
+  justify(some(inits([
     context({ syntax: { inline: {
       label: false,
     }}},